Factor out `Display` impl for vectors
