+++ ext/crates/fp/src/vector/impl_fqslice.rs
@@ -1,5 +1,3 @@
-use itertools::Itertools;
-
 use super::inner::{FqSlice, FqVector};
 use crate::field::{Field, element::FieldElement};
 
@@ -43,31 +41,5 @@ impl<'a, F: Field> FqSlice<'a, F> {
 impl<'a, F: Field> From<&'a FqVector<F>> for FqSlice<'a, F> {
     fn from(v: &'a FqVector<F>) -> Self {
         v.slice(0, v.len())
-    }
-}
-
-impl<F: Field> std::fmt::Display for FqSlice<'_, F> {
-    /// # Example
-    /// ```
-    /// # use fp::field::{Field, SmallFq};
-    /// # use fp::prime::{P2, ValidPrime};
-    /// # use fp::vector::FqVector;
-    /// let fq = SmallFq::new(P2, 3);
-    /// let v = FqVector::from_slice(fq, &[fq.zero(), fq.one(), fq.a(), fq.a() * fq.a()]);
-    /// assert_eq!(&format!("{v}"), "[0, 1, a, a^2]");
-    ///
-    /// // This only looks reasonable over prime fields of order less than 10
-    /// assert_eq!(&format!("{v:#}"), "01aa^2");
-    /// ```
-    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
-        if f.alternate() {
-            for v in self.iter() {
-                // If self.p >= 11, this will look funky
-                write!(f, "{v}")?;
-            }
-            Ok(())
-        } else {
-            write!(f, "[{}]", self.iter().format(", "))
-        }
     }
 }
